move existing animation tests to animation plugin tests
